Major rewrite
Instead of the server and client being distinct and separate, the server class
includes the client automatically. This is in part due to how some OS insist on
managing both components with a single initscript and/or use configuration
files shared between them.

Because of this we can now also support RHEL/CentOS 6 and Ubuntu 16.04.